Refactor the tests into functional form
