--- test/definitions.jl
@@ -1,46 +1,0 @@
-function _conv(h, x)
-    m = length(h)
-    n = length(x)
-    zero_pad_x(i) = 1 <= i <= n ? x[i] : 0
-    return [sum(h[j] * zero_pad_x(i - j + 1) for j in 1:m) for i in 1:m+n-1]
-end
-
-@testset "Conv (issue #364)" begin
-    n = 3
-    m = 11
-    h = rand(m)
-    x = rand(n)
-    hvar = Variable(m)
-    hvar.value = h
-    @test evaluate(conv(hvar, x)) ≈ _conv(h, x)
-end
-
-@testset "`conj` (issue #416)" begin
-    A = [1 1im; -1im 1]
-    X = ComplexVariable(2, 2)
-    p = minimize(real(tr(conj(X))), [X == A])
-    solve!(
-        p,
-        MOI.OptimizerWithAttributes(
-            SCS.Optimizer,
-            "verbose" => 1,
-            "eps_abs" => 1e-6,
-        ),
-    )
-    @test evaluate(X) ≈ A atol = 1e-3
-end
-
-@testset "`logisticloss` (issue #458)" begin
-    x = Variable()
-    expr = logisticloss(x)
-    @test expr isa Convex.AbstractExpr
-
-    set_value!(x, 1.5)
-    @test evaluate(expr) ≈ log(1 + exp(1.5))
-end
-
-@testset "`dot` (issue #508)" begin
-    x = [1.0 + 1.0im]
-    y = [-1.0im]
-    @test dot(x, y) ≈ evaluate(dot(constant(x), y))
-end
